Player movement fixes

+++ src/room.rs
@@ -49,7 +49,7 @@ impl Room {
 
     pub fn tile_at_index(&self, x: u32, y: u32) -> Tile {
         let kind = *self.tiles
-            .get((self.width * y) as usize + x as usize)
+            .get(self.width as usize * y as usize + x as usize)
             .unwrap_or(&TileKind::Empty);
         let rect = Rect::new(
             (x * self.tile_size) as f32,
